Adds a procedure for remote queries.
For now, the queries are only allowed for one model at a time, and the
filtering is only by equality and it's clauses are joined by 'AND'.

Also, since values given to the filters are currently passed as URL
parameters, they can only be strings or numbers for them to work
properly.