Corrected (hopefully) a serious flaw in operation compression.
The `compressed_operations' procedure was acting like the mutable
version `compress' in that it assumed the sequences were local-like,
that is, no primary key was being reused after it was expended. Since
the procedure `compressed_operations' acts on the pull message, and
the message contains operations from a multitude of nodes, primary key
reuse *is* possible.

For example, the sequence insert-delete can happen in a single
node. Another node can then reinsert the record during conflict
resolution, to form the sequence insert-delete-insert. The previous
version of `compressed_operations' discarded such a sequence and no
operation was performed.