Tracking directions to exclude models from synchronization.
The default direction is ('push', 'pull'). This means a model's
operations are included in both push and pull messages. Giving only
one of them means the other procedure's messages will exclude those
operations and objects. There's no option to give no direction since
that's roughly equivalent to not tracking the class at all.

This can be used to have node-computed tables that get forwarded to
the server but aren't part of the merge subroutine ('push' only), or
to have server-computed tables that function conversely.